Remove unnecessary and incorrect cast.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@53287 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/dbapi/driver/ftds64/freetds/tds/net.c
@@ -420,7 +420,7 @@ tds_goodread(TDSSOCKET * tds, unsigned char *buf, int buflen, unsigned char unfi
 		got += len;
 
 		now = GetTimeMark();
-		if (tds->query_timeout > 0 && (int)now - start >= tds->query_timeout) {
+		if (tds->query_timeout > 0 && now - start >= tds->query_timeout) {
 
 			int timeout_action = TDS_INT_CONTINUE;
